use int rather than size_t to avoid printf() warnings in common code
git-svn-id: svn://localhost/svn/u-boot/trunk@1524 e8ad1c53-384c-47db-b5ce-66ae1f620dea

+++ u-boot-2008.10/cpu/blackfin/traps.c
@@ -261,7 +261,7 @@ static char *strexcause(uint16_t excause)
 void dump(struct pt_regs *fp)
 {
 	char buf[150];
-	size_t i;
+	int i;
 	uint16_t hwerrcause, excause;
 
 	if (!ENABLE_DUMP)
@@ -337,7 +337,7 @@ void dump_bfin_trace_buffer(void)
 {
 	char buf[150];
 	unsigned long tflags;
-	size_t i = 0;
+	int i = 0;
 
 	if (!ENABLE_DUMP)
 		return;
